+++ src/Controller/view_appointments.java
@@ -17,6 +17,7 @@ import java.io.IOException;
 import java.net.URL;
 import java.sql.PreparedStatement;
 import java.sql.SQLException;
+import java.util.Optional;
 import java.util.ResourceBundle;
 
 import static Model.JDBC.createCustomerList;
@@ -109,24 +110,30 @@ public class view_appointments implements Initializable {
             Alert alert = new Alert(Alert.AlertType.CONFIRMATION);
             alert.setTitle("Delete Appointment?");
             alert.setContentText("Are you sure?");
-            alert.showAndWait();
-            Appointment a = appointmentTable.getSelectionModel().getSelectedItem();
-
-            String sql = "DELETE FROM appointments WHERE Appointment_ID = ?";
-            PreparedStatement ps = JDBC.connection.prepareStatement(sql);
-            ps.setInt(1, a.getApptID());
-            int rowsAffected = ps.executeUpdate();
-            if (rowsAffected > 0)
-                System.out.println(rowsAffected + " appointment deleted from table");
-            else {
-                Alert alert1 = new Alert(Alert.AlertType.ERROR);
-                alert1.setTitle("Deletion Error:");
-                alert1.setHeaderText("Appointment not found");
-                alert1.showAndWait();
+            Optional<ButtonType> result = alert.showAndWait();
+            if (result.isPresent() && result.get() == ButtonType.OK) {
+                Appointment a = appointmentTable.getSelectionModel().getSelectedItem();
+                String sql = "DELETE FROM appointments WHERE Appointment_ID = ?";
+                PreparedStatement ps = JDBC.connection.prepareStatement(sql);
+                ps.setInt(1, a.getApptID());
+                int rowsAffected = ps.executeUpdate();
+                if (rowsAffected > 0) {
+                    Alert newAlert = new Alert(Alert.AlertType.INFORMATION);
+                    newAlert.setTitle("Success");
+                    newAlert.setHeaderText("Appointment Cancelled");
+                    newAlert.setContentText("Appointment successfully deleted from DB");
+                    newAlert.showAndWait();
+                }
+                else {
+                    Alert alert1 = new Alert(Alert.AlertType.ERROR);
+                    alert1.setTitle("Deletion Error:");
+                    alert1.setHeaderText("Appointment not found");
+                    alert1.showAndWait();
+                }
+                appointmentTable.getSelectionModel().clearSelection();
+                appointmentTable.getItems().clear();
+                appointmentTable.setItems(JDBC.createAppointmentList());
             }
-            appointmentTable.getSelectionModel().clearSelection();
-            appointmentTable.getItems().clear();
-            appointmentTable.setItems(JDBC.createAppointmentList());
         }
     }
